backend/requirements/base.txt
django==4.2.16
djangorestframework==3.15.2
django-environ==0.11.2
django-cors-headers==4.4.0
django-extensions==3.2.3
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
celery==5.4.0
redis==5.1.1
psycopg2-binary==2.9.9
kubernetes==30.1.0
# langchain-ollama 0.2.0 requires langchain-core>=0.3.0; use langchain 0.3.x
langchain-core>=0.3.0,<0.4
langchain>=0.3.0,<0.4
langgraph==0.2.28
langchain-ollama==0.2.0
langchain-chroma>=0.1.4
chromadb==0.5.15
neo4j==5.24.0
mcp==1.1.0
pydantic==2.8.2
